Speed-up on 600 PbPb events generated by @davidrohr :+1: 
dev:
[19467:tof-matcher]: [16:51:08][INFO] Timing Do Matching: Cpu: 3.776e+01 s Real: 3.772e+01 s in 0 slots
[19467:tof-matcher]: [16:51:08][INFO] Timing Do Matching ITSTPC: Cpu: 0.000e+00 s Real: 0.000e+00 s in 0 slots
[19467:tof-matcher]: [16:51:08][INFO] Timing Do Matching TPC   : Cpu: 3.755e+01 s Real: 3.751e+01 s in 17 slots
[19467:tof-matcher]: [16:51:08][INFO] Shifting Z for 334358 matched TPC tracks according to TOF time info
[19468:TOFCalibWriter]: [16:51:09][INFO] RECEIVED MATCHED SIZE 334358
[19467:tof-matcher]: [16:51:09][INFO] TOF matching total timing: Cpu: 5.215e+01 Real: 5.211e+01 s in 1 slots
[19469:TOFMatchedWriter_TPC]: [16:51:09][INFO] RECEIVED MATCHED SIZE 334358

this branch:
[5487:tof-matcher]: [16:42:51][INFO] Timing Do Matching: Cpu: 2.891e+01 s Real: 2.888e+01 s in 0 slots
[5487:tof-matcher]: [16:42:51][INFO] Timing Do Matching ITSTPC: Cpu: 0.000e+00 s Real: 0.000e+00 s in 0 slots
[5487:tof-matcher]: [16:42:51][INFO] Timing Do Matching TPC   : Cpu: 2.870e+01 s Real: 2.866e+01 s in 17 slots
[5487:tof-matcher]: [16:42:51][INFO] Shifting Z for 334358 matched TPC tracks according to TOF time info
[5488:TOFCalibWriter]: [16:42:51][INFO] RECEIVED MATCHED SIZE 334358
[5487:tof-matcher]: [16:42:51][INFO] TOF matching total timing: Cpu: 4.460e+01 Real: 4.456e+01 s in 1 slots

==> ~25%

Compared to the initial measurements (~536s) ==> we get to a factor 18 improvement. 
@noferini , please check that you "like" the changes, of course.