@noferini 2021-09-16
Hi Chiara,
I checked your changes. They look fine.
For my curiosity, do you also have also the change of the performance when matching simultaneously TPC and ITS-TPC tracks?