@chiarazampolli 2021-09-17
Further improvement added as discussed with @noferini , like this the o2::tof::Geo::getSector() will be called much less. I will measure in detail next week.