@chiarazampolli 2021-09-22
Hello @noferini , @shahor02 , @davidrohr , 

With current dev:

*TPC only:
[143076:tof-matcher]: [22:36:30][INFO] Timing Do Matching:        Cpu: 2.813e+01 s Real: 2.810e+01 s in 0 slots
[143076:tof-matcher]: [22:36:30][INFO] Timing Do Matching ITSTPC: Cpu: 0.000e+00 s Real: 0.000e+00 s in 0 slots
[143076:tof-matcher]: [22:36:30][INFO] Timing Do Matching TPC   : Cpu: 2.787e+01 s Real: 2.786e+01 s in 17 slots
[143076:tof-matcher]: [22:36:30][INFO] Shifting Z for 334358 matched TPC tracks according to TOF time info
[143077:TOFCalibWriter]: [22:36:30][INFO] RECEIVED MATCHED SIZE 334358
[143076:tof-matcher]: [22:36:30][INFO] TOF matching total timing: Cpu: 4.375e+01 Real: 4.371e+01 s in 1 slots


*ITSTPC and TPC:
[143447:tof-matcher]: [22:39:35][INFO] Timing Do Matching:        Cpu: 2.170e+01 s Real: 2.168e+01 s in 0 slots
[143447:tof-matcher]: [22:39:35][INFO] Timing Do Matching ITSTPC: Cpu: 3.520e+00 s Real: 3.498e+00 s in 17 slots
[143447:tof-matcher]: [22:39:35][INFO] Timing Do Matching TPC   : Cpu: 1.802e+01 s Real: 1.802e+01 s in 17 slots
[143447:tof-matcher]: [22:39:35][INFO] Shifting Z for 213128 matched TPC tracks according to TOF time info
[143448:TOFCalibWriter]: [22:39:35][INFO] RECEIVED MATCHED SIZE 318549
[143449:TOFMatchedWriter]: [22:39:35][INFO] RECEIVED MATCHED SIZE 105421
[143450:TOFMatchedWriter_TPC]: [22:39:35][INFO] RECEIVED MATCHED SIZE 213128
[143447:tof-matcher]: [22:39:35][INFO] TOF matching total timing: Cpu: 3.608e+01 Real: 3.604e+01 s in 1 slots

So, not much after the very last commit, but everything helps. 

Chiara